Extract all require/load operations to a `Loader` class
